Fix resource_path error on project startup
When a GOAT project opens, there is an error message printed in the
"Output" tab:

```
Can't save resource to empty path. Provide non-empty path or a Resource with non-empty resource_path.
```

This was caused by the fact that the default audio bus file was removed
from Project Settings.

This commit adds the file back, and ensures that GOAT audio buses will
be added to it only once.